Commit 5 - Fixed Crud Operation in CarService deleteSingleCar: if the id is not in the db the response will have a Conflict HTTP status

+++ src/main/java/kemy/dev/CrudApiExample/services/CarService.java
@@ -4,14 +4,14 @@ import kemy.dev.CrudApiExample.entities.Car;
 import kemy.dev.CrudApiExample.repositories.CarRepository;
 import lombok.Data;
 import lombok.ToString;
-import net.bytebuddy.dynamic.DynamicType;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
-
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
 
@@ -21,8 +21,16 @@ import java.util.Optional;
 public class CarService {
     @Autowired
     CarRepository carRepository;
+
+// to run getSingleCar() you must create this attribute
     @Autowired
     Optional <Car> car ;
+
+// to run deleteSingleCar() you must create these attributes:
+    @Autowired
+    HttpServletRequest request;
+    @Autowired
+    HttpServletResponse response;
 
 
     //CREATE
@@ -33,6 +41,7 @@ public class CarService {
     public List<Car> getAllCars(){
         return carRepository.findAll();
     }
+    // gets a specific Car - if absent, the response will be an empty Car
     public Optional<Car> getSingleCar(@PathVariable long id){
 
         if(carRepository.existsById(id)){
@@ -45,7 +54,7 @@ public class CarService {
         }
     }
     //UPDATE
-
+    // edit a specific Car - if absent, it will create an empty Car
    public Car editCar (@PathVariable long id,@RequestBody Car car) {
        if (carRepository.existsById(id)) {
        car.setId(id);
@@ -55,13 +64,25 @@ public class CarService {
        return carRepository.saveAndFlush(new Car());
     }
 
+   //DELETE
 
-    //DELETE
-    public String deleteCar(@PathVariable long id){
-        carRepository.deleteById(id);
-        System.out.println("deleted car id: "+id);
-        return "deleted car ( id: "+id+")";
+   // deletes a specific Car - if absent, the response will have a Conflict HTTP status
+   public String deleteSingleCar(@PathVariable long id) {
+
+        if (!carRepository.existsById(id)) {
+            System.out.println(HttpStatus.CONFLICT);
+            try {
+                response.sendError(409, "CONFLICT");
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+        } else
+            carRepository.deleteById(id);
+            System.out.println("car with  id " + id + " was deleted");
+            return "car with  id " + id + " was deleted";
+
     }
+
     public String deleteAllCars(){
         carRepository.deleteAll();
         System.out.println("deleted all cars");
